refactor: extract scroll prevention logic to shared hook
